+++ CHANGELOG.md
@@ -105,3 +105,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Filter output formats for DataSHIELD
 - Flag ENABLE_TOKEN_MANAGER
 - Content Disposition in exposed headers
+- Exists Votum and Application Form